Tests for Skill Service and Controller (#31)
* Remove MediaType acceptance from /admin/users API test

* Add Mock Methods for Skill and Talent in Test Util

* Add Skill Service tests

Added JUnit tests for the SkillService class in the AccountService. The tests cover various scenarios such as fetching all skills, updating a talent's skills, and handling exceptions when a requested talent or skill is not found.

* Added Builder annotation to SkillResponse model

* Add SkillControllerTest class for unit testing

A new file SkillControllerTest.java has been added in the test directory. The file includes unit tests for SkillController, mainly focusing on testing the "get all skills" API endpoint.