Tests for Report Service (#32)
* Add new mock entities in MockModelsUtil

* Add ReportServiceTest class for testing report service

This commit introduces the ReportServiceTest class. This class is essential for conducting unit tests on the report service functionality. It includes tests for successfully reporting a talent or sponsor, handling scenarios when talent or sponsor is not found, and checking text toxicity to decide whether to report a talent or sponsor.

* Add tests for SponsorController

This commit introduces SponsorControllerTest.java, where two unit tests are implemented. The first tests the successful reporting of a sponsor to ensure that the function performs correctly in normal circumstances. The second test handles the scenario where a sponsor is not found to verify that the system throws an appropriate exception in such cases.

* Add TalentController tests in AccountService

A new file, TalentControllerTest.java, was created to add comprehensive tests for TalentController. The unit tests cover two main scenarios: successful talent reporting and handling talent not found exception. expected.

* Removed unnecessary imports